requirements.txt
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.1.5
attrs==18.1.0
autoenv==1.0.0
autopep8==1.3.5
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.3.9
coverage==4.5.1
coveralls==1.3.0
Django==2.1.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.7
django-crispy-forms==1.7.2
django-multiselectfield==0.1.8
docopt==0.6.2
flasgger==0.9.2
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-JWT-Extended==3.13.1
gunicorn==19.8.1
heroku==0.1.4
idna==2.7
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.4
more-itertools==4.2.0
Pillow==5.4.1
pipenv==2018.7.1
pluggy==0.6.0
psycopg2==2.7.6.1
py==1.5.3
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.6.2
pytest-cov==2.6.0
python-dateutil==1.5
python-dotenv==0.9.1
python-env==1.0.0
pytz==2018.4
PyYAML==3.13
requests==2.21.0
six==1.11.0
stripe==2.19.0
typed-ast==1.1.0
typing==3.6.6
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
wrapt==1.10.11
